# spec/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/batches/{batchId}:
    get:
      operationId: BatchIdController_getOne_v1
      parameters:
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Get growing batch by id
      tags:
        - batches
    patch:
      operationId: BatchIdController_update_v1
      parameters:
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBatchDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '412':
          description: >-
            Precondition Failed: Cannot modify start_at if batch has started, or cannot modify fields other than
            description/notes if batch has finished
          content:
            application/json:
              schema:
                example:
                  statusCode: 412
                  error: Precondition failed
                  message: The server's expectations aren't being met
                  emitter: Middleware
                  timestamp: '2024-09-14T22:00:00.000Z'
                  path: /path/of/the/failing/route
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Modify an existing batch
      tags:
        - batches
    delete:
      operationId: BatchIdController_remove_v1
      parameters:
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      responses:
        '204':
          description: Deleted
        '404':
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Delete a batch
      tags:
        - batches
...
  /v1/batches/{batchId}/recipe:
    post:
      description: >-
        Creates a new recipe using the species and climate targets from the batch. duration_days is computed from
        start_at → finish_at. The batch must be finished and have species, temperature_target, and humidity_target set.
      operationId: BatchIdRecipeController_createRecipe_v1
      parameters:
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeFromBatchDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Create a recipe from a finished batch
      tags:
        - batches
...
info:
  title: mushpi-server
  description: A backend for the Mushroom Pi project
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    secret:
      type: http
      scheme: bearer
      in: header
      name: authorization
  schemas:
...
    ErrorDto:
      type: object
      properties: {}
...
    Batch:
      type: object
      properties:
        images:
          description: Filenames of attached images
          type: array
          items:
            type: string
        images_url:
          description: Absolute URLs to the batch images (computed)
          example:
            - http://localhost:3000/images/batches/7/1.jpg
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - planned
            - in-progress
            - finished
          description: Current status of the batch (computed from start_at and finish_at)
          example: in-progress
        images_left:
          type: number
          description: Number of images that can still be uploaded to this batch
          example: 3
          minimum: 0
          maximum: 5
      required:
        - status
        - images_left
...
    UpdateBatchDto:
      type: object
      properties:
        start_at:
          type: string
        finish_at:
          type: string
        species:
          type: string
        temperature_target:
          type: number
          minimum: 0
          maximum: 50
        humidity_target:
          type: number
          minimum: 20
          maximum: 90
        notes:
          type: string
          maxLength: 2000
        description:
          type: string
          maxLength: 120
    CreateRecipeFromBatchDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        notes:
          type: string
          maxLength: 2000
      required:
        - name
    Recipe:
      type: object
      properties:
        image:
          type: string
          description: Filename of the uploaded image or external URL
          example: 1.jpg
        image_url:
          type: string
          nullable: true
          description: Absolute URL to the image
          example: http://localhost:3000/images/recipes/1.jpg
      required:
        - image_url
...